trying lowp

+++ src/shaders.cpp
@@ -300,9 +300,10 @@ void AnyAminoShader::drawElements(GLushort *indices, GLsizei elements, GLenum mo
 ColorShader::ColorShader() : AnyAminoShader() {
     //shaders
     //Note: no performance difference seen between highp, mediump and lowp!
+    //cbxx trying lowp instead of highp
     fragmentShader = R"(
 #ifdef EGL_GBM
-        precision highp float;
+        precision lowp float;
 #endif
         uniform vec4 color;
